tests: update tests with last refactor
